requirements.txt
flask
flask-cors
transformers
torch
Pillow
sqlalchemy
streamlit
requests
python-dotenv
